Header and Front Page
I've added a front page for after you click the submit button on the login page. The front page has a header, which has been given its own file so it can be included in future pages. The code is quite messy, and the format isn't perfect but it works for now.